refactor(efd-server): ♻️ update town meeting hall transitions
